Partially implement proper flash storage.

+++ stm/malloc0.c
@@ -6,7 +6,7 @@ static uint32_t mem = 0;
 void *malloc(size_t n) {
     if (mem == 0) {
         extern uint32_t _heap_start;
-        mem = &_heap_start; // need to use big ram block so we can execute code from it (is it true that we can't execute from CCM?)
+        mem = (uint32_t)&_heap_start; // need to use big ram block so we can execute code from it (is it true that we can't execute from CCM?)
     }
     void *ptr = (void*)mem;
     mem = (mem + n + 3) & (~3);
